+++ components/index.js
@@ -43,6 +43,7 @@ export * from './tbt-dropdown.js';
 export * from './tbt-multiselect.js';
 export * from './tbt-datepicker.js';
 export * from './tbt-search.js';
+export * from './tbt-color-picker.js';
 export * from './tbt-form.js';
 export * from './tbt-address.js';